+++ hdeeprm/manager.py
@@ -97,6 +97,7 @@ Attributes:
         self.state_changes = {}
         with open('./res_hierarchy.pkl', 'rb') as in_f:
             self.platform, self.core_pool = pickle.load(in_f)
+        logging.info('Reference speed %s', self.platform['reference_speed'])
         # Add the job resource requirement limits to the resource hierarchy
         with open('./job_limits.pkl', 'rb') as in_f:
             self.platform['job_limits'] = pickle.load(in_f)